The DeepGo Project
The DeepGo Project has been created in Java. The IDE used is netbeans. The project is a working project but there are a few issues with the java database connectivity to mysql , where the user login authentication is concerned. Otherwise, the project runs fine, takes queries and returns articles regarding the keyword from Wikipedia. The DeepGo Class is the class with the main function. Do remember to change the file path of the images used in the application. 

PS.
Do make sure you have the following installed:
jsoup library
Java JDBC for mysql 
MySQL database
JavaFX library